Fix RouteBase string formatting bug
When correcting conversion issues in RouteBase a to_string bug was
introduced which caused some characters to be escaped when inserted
to the string, for example a prefix_len of 0 would render as "\0"
rather than inserting '0'. The std::ios::binary flag does not seem
to prevent this for std::ostringstream so I have cast the data member
up to uint16_t which should be safe, and solves the issue.

Added a unit test to demonstrate the issue. Old code output was
"0.0.0.0/\0", now outputs "0.0.0.0/0" as expected.

+++ openvpn/tun/builder/capture.hpp
@@ -150,7 +150,7 @@ class TunBuilderCapture : public TunBuilderBase, public RC<thread_unsafe_refcoun
         std::string to_string() const
         {
             std::ostringstream os;
-            os << address << '/' << prefix_length;
+            os << address << '/' << static_cast<uint16_t>(prefix_length);
             if (!gateway.empty())
                 os << " -> " << gateway;
             if (metric >= 0)
